+++ actix-tls/src/connect/uri.rs
@@ -1,8 +1,21 @@
-use http::Uri;
+use http as http_02;
 
 use super::Host;
 
-impl Host for Uri {
+impl Host for http_02::Uri {
+    fn hostname(&self) -> &str {
+        self.host().unwrap_or("")
+    }
+
+    fn port(&self) -> Option<u16> {
+        match self.port_u16() {
+            Some(port) => Some(port),
+            None => scheme_to_port(self.scheme_str()),
+        }
+    }
+}
+
+impl Host for http_1::Uri {
     fn hostname(&self) -> &str {
         self.host().unwrap_or("")
     }