+++ actix-tls/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Added support to `http` crate version `1.0`.
+
 ## 3.1.1
 
 - Fix `rustls` v0.21 version requirement.