+++ actix-tls/CHANGES.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Added support to `http` crate version `1.0`.
+- Add `rustls-0_21-native-roots` and `rustls-0_20-native-roots` crate features which utilize the `rustls-native-certs` crate to enable a `native_roots_cert_store()` functions in each rustls-based `connect` module.
+- Implement `Host` for `http::Uri` (`http` crate version `1`).
 
 ## 3.1.1